+++ jsonargparse/actions.py
@@ -155,7 +155,7 @@ class ActionConfigFile(Action, FilesCompleterMethod):
     def __init__(self, **kwargs):
         """Initializer for ActionConfigFile instance."""
         if 'default' in kwargs:
-            raise ValueError('default not allowed for ActionConfigFile, use default_config_files.')
+            self.set_default_error()
         opt_name = kwargs['option_strings']
         opt_name = opt_name[0] if len(opt_name) == 1 else [x for x in opt_name if x[0:2] == '--'][0]
         if '.' in opt_name:
@@ -171,6 +171,10 @@ class ActionConfigFile(Action, FilesCompleterMethod):
             TypeError: If there are problems parsing the configuration.
         """
         self.apply_config(parser, cfg, self.dest, values)
+
+    @staticmethod
+    def set_default_error():
+        raise ValueError('ActionConfigFile does not accept a default, use default_config_files.')
 
     @staticmethod
     def apply_config(parser, cfg, dest, value) -> None: